Clean up SQL database related functions
