+++ internal/sqldb/sqlite/sqlite.go
@@ -20,26 +20,15 @@ import (
 	"path/filepath"
 )
 
-func CreateDB(fileDir string) (*sql.DB, error) {
-	// Create all intermediate directories.
-	if err := os.MkdirAll(fileDir, 0755); err != nil {
-		return nil, err
-	}
+func ConnectDB(fileDir string) (*sql.DB, error) {
 	dbPath := filepath.Join(fileDir, "datacommons.db")
 	_, err := os.Stat(dbPath)
-	if err == nil {
-		sqlClient, err := sql.Open("sqlite3", dbPath)
-		if err != nil {
-			return nil, err
-		}
-		return sqlClient, nil
-	}
-	if !os.IsNotExist(err) {
-		return nil, err
-	}
-	_, err = os.Create(dbPath)
 	if err != nil {
 		return nil, err
 	}
-	return sql.Open("sqlite3", dbPath)
+	sqlClient, err := sql.Open("sqlite3", dbPath)
+	if err != nil {
+		return nil, err
+	}
+	return sqlClient, nil
 }
