+++ internal/sqldb/sqlquery/check_svg.go
@@ -12,31 +12,32 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package query
+package sqlquery
 
 import (
 	"database/sql"
+	"fmt"
 
-	pb "github.com/datacommonsorg/mixer/internal/proto"
+	"github.com/datacommonsorg/mixer/internal/util"
 )
 
-// GetProvenances returns all the provenance name and url in SQL database.
-func GetProvenances(sqlClient *sql.DB) (map[string]*pb.Facet, error) {
-	result := map[string]*pb.Facet{}
-	query :=
+// CheckVariableGroups check and returns variable groups that in SQL database.
+func CheckVariableGroups(sqlClient *sql.DB, variableGroups []string) ([]string, error) {
+	result := []string{}
+	// Find all the sv that are in the sqlite database
+	query := fmt.Sprintf(
 		`
-			SELECT t1.subject_id, t2.object_value, t3.object_value
-			FROM triples AS t1
-			JOIN triples AS t2 ON t1.subject_id = t2.subject_id
-			JOIN triples AS t3 ON t1.subject_id = t3.subject_id
-			WHERE t1.predicate = "typeOf"
-			AND t1.object_id = "Provenance"
-			AND t2.predicate = "name"
-			AND t3.predicate = "url"
-		`
+			SELECT DISTINCT(subject_id) FROM triples
+			WHERE predicate = "typeOf"
+			AND subject_id IN (%s)
+			AND object_id = 'StatVarGroup';
+		`,
+		util.SQLInParam(len(variableGroups)),
+	)
 	// Execute query
 	rows, err := sqlClient.Query(
 		query,
+		util.ConvertArgs(variableGroups)...,
 	)
 	if err != nil {
 		return nil, err
@@ -44,15 +45,12 @@ func GetProvenances(sqlClient *sql.DB) (map[string]*pb.Facet, error) {
 	defer rows.Close()
 	// Process the query result
 	for rows.Next() {
-		var id, name, url string
-		err = rows.Scan(&id, &name, &url)
+		var svg string
+		err = rows.Scan(&svg)
 		if err != nil {
 			return nil, err
 		}
-		result[id] = &pb.Facet{
-			ImportName:    name,
-			ProvenanceUrl: url,
-		}
+		result = append(result, svg)
 	}
 	return result, nil
 }
