+++ internal/sqldb/sqlquery/count_observation.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package query
+package sqlquery
 
 import (
 	"database/sql"
