--- internal/sqldb/test_data/geo_id.csv
@@ -1,3 +1,0 @@
-geoId,observationDate,Count_Person,Count_Person_Male
-06,2021,555,666
-06085,2022,333,444
