--- internal/sqldb/test_data/lat_lng.csv
@@ -1,2 +1,0 @@
-lat#lng,observationDate,Count_Person,Count_Person_Male
-38.7#-119.4,2021,555,666
