--- internal/sqldb/test_data/name.csv
@@ -1,2 +1,0 @@
-name,observationDate,Count_Person,Count_Person_Male
-Beijing,2021,555,666
